requirements.txt
py-clob-client>=0.34
requests>=2.28
python-dotenv>=1.0
websockets>=12.0
pytz>=2024.1
aiohttp>=3.9
orjson>=3.9
aiosqlite>=0.19
rich>=13.7
